pyproject.toml
[tool.poetry]
name = "productdescriber"
version = "0.1.1"
description = "AI ProductDescriber"
authors = ["katiridium"]
readme = "README.md"
packages = [
  { include = "productdescriber" },
]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
python-dotenv = "^1.1.0"
jinja2 = "^3.1.6"
aiofiles = "^24.1.0"
python-multipart = "^0.0.20"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.0"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"
flake8 = "^7.2.0"
coverage = "^7.8.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
